Update to v1.2.0
Addition of switch_screens(options) function for developers to use in
their custom JS code.

Addition of scrolling to the top of the page before animating (more
smooth than snapping to the top). Time can be changed with
data-ms-scroll-time, and default can be set with set_defaults(options)
and set_default_time(time, type) with type ‘scroll’